+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 1.2.0
 * Move project to Arch Loom
+* Add rotated block support
+* Add example mushroom features
 
 # 1.1.2
 * Fix Tree Decorator Types not being registered on Fabric